Fixed: java.lang.Long cannot be cast to java.lang.Integer

+++ app/models/IngredientSupply.java
@@ -73,7 +73,7 @@ public class IngredientSupply extends JpaModel {
 			int size = Play.application().configuration().getInt("page.size");
 			if (index < 1)
 				index = 1;
-			Integer rows = (Integer) JPA.em()
+			Long rows = (Long) JPA.em()
 					.createQuery("select count(*) from IngredientSupply")
 					.getSingleResult();
 			List<IngredientSupply> list = JPA.em()
